pack-write: refactor renaming in finish_tmp_packfile()
Refactor the renaming in finish_tmp_packfile() into a helper
function. The callers are now expected to pass a "name_buffer" ending
in "pack-OID." instead of the previous "pack-", we then append "pack",
"idx" or "rev" to it.

By doing the strbuf_setlen() in rename_tmp_packfile() we re-use the
buffer and avoid the repeated allocations we'd get if that function
had its own temporary "struct strbuf".

This approach of re-using the buffer does make the last user in
pack-object.c's write_pack_file() slightly awkward, we needlessly do a
strbuf_setlen() there just before the strbuf_release() for
consistency. In subsequent changes we'll move that bitmap writing code
around, so let's not skip the strbuf_setlen() now.

The previous strbuf_reset() idiom originated with
5889271114a (finish_tmp_packfile():use strbuf for pathname
construction, 2014-03-03), which in turn was a minimal adjustment of
pre-strbuf code added in 0e990530ae (finish_tmp_packfile(): a helper
function, 2011-10-28).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>