credential-cache{,--daemon}: don't build under NO_UNIX_SOCKETS
Change the implementation of b5dd96b70ac (make credential helpers
builtins, 2020-08-13) to not build these at all under
NO_UNIX_SOCKETS.

This is the easiest way to get rid of one out of two users of an
obscure parse_options() API I'm trying to get rid of. It does mean
that the goal of emitting a custom error message in b5dd96b70ac is
being eliminated, but per [1] that seems to be an OK direction to go
in.

By not compiling it at all it won't be included in the "struct
cmd_struct", and therefore will also be omitted from
"--list-cmds=builtins".

1. https://lore.kernel.org/git/cover-v2-0.6-00000000000-20210910T153146Z-avarab@gmail.com/

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>