git.c: add a NEED_UNIX_SOCKETS option for built-ins
Change the implementation of b5dd96b70ac (make credential helpers
builtins, 2020-08-13) to declare in the "struct cmd_struct" that
NO_UNIX_SOCKETS can't be set.

This is one of two in-tree users for the empty lines in
parse_options() usage, getting rid of that is the main motivation for
this, but it also doesn't make sense to emit these sorts of usage
messages just to appease t0012-help.sh, which seemingly b5dd96b70ac
aimed to do.

I.e. these commands don't support "[options]", or "<action>" so
emitting that at the beginning is incorrect. We should just die right
away.

The existing code also had the edge case of not emitting the die()
message if a "-h" argument was given, since parse_options() will
handle the exit() itself in that case. We could feed it
PARSE_OPT_NO_INTERNAL_HELP, but this is better.

By moving this up to the "struct cmd_struct" we can also skip these in
--list-cmds=builtins instead, as noted above we shouldn't be exiting
with code 129 in these cases.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>