format-patch: increase --range-diff's --creation-factor to "100"
Change the default "format-patch" --creation-factor which is used with
the --range-diff option to 100, from the previous default of 60, which
was the same as "range-diff" itself.

It's been my experience that increasing this is more helpful than not
when re-rolling a series of patches, and the Git ML occasionally gets
a range-diff as part of a cover letter that's only adding and removing
commits, since we punted on the similarity detection.

We can now remove the RANGE_DIFF_CREATION_FACTOR_DEFAULT macro, as the
reasons for adding it in 25668659bfd (range-diff: publish default
creation factor, 2018-07-22) no longer apply.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>