difftool: use "struct strvec" API in run_{dir,file}_diff()
The underlying run_command() API can take either the "struct strvec
args", or a "const char **argv". Let's move to the former to use the
more "native" version of run_command() in both of these functions.

This change probably isn't worth in on its own, but sets us up to
simplify API use even more in a subsequent commit.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>