Makefile: define LIB_H files inline, again
Before [1] the LIB_H list used to be static, but was then replaced
with a "find" command. Then in [2] we started using "git ls-files",
and some time later fixed some bugs in it [3].

The reason I'd like to move back to exhaustively listing the LIB_H
files is to make the Makefile deterministic and easy to understand,
and so that our known list of *.c, *.h etc. files can be fed to
targets such as the "TAGS" target.

We need to make sure that in doing so we're not bringing back the
problem of the list potentially drifting out of sync, we're doing that
by asserting that we're not getting any unknown headers.

Let's move this back to a hardcoded list. This solves the problem of
over-supplying dependencies, as well as making the Makefile run faster
by not relying on a "find" command. It'll also bring back the problem
of this list potentially drifting out of sync.

1. d85b0dff729 (Makefile: use `find` to determine static header
   dependencies, 2014-08-25)
2. 92b88eba9f7 (Makefile: use `git ls-files` to list header files, if
   possible, 2019-03-04)
3. 604a64641d8 (Makefile: dedup list of files obtained from ls-files,
   2019-04-21)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>