Makefile: split LIB_COMPAT_OBJS off from LIB_OBJS
In a subsequent commit we'll start asserting that variables like
LIB_OBJS are sorted, to do that we need to split off the
LIB_COMPAT_OBJS. This is because we'll assign to it after the
main (sorted) assignment.

I think this also makes the flow easier to follow as it's obvious
later on which 3rd party library dependencies come from compat/ and
which don't. Note that there is an existing COMPAT_OBJS, this is
different, these are things in compat/* we treat as though they were
LIB_OBJS itself.

We need to be careful not to brake the make infrastructure added in
061c2240b1b (Introduce CMake support for configuring Git, 2020-06-12).

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>