Makefile: split BUILT_INS_EXTRA off from BUILT_INS
In preparation for checking the sorting of hardcoded lists in the
Makefile let's split up the part of the BUILT_INS that are manually
defined into its own variable.

Let's move a comment added in 1736855c9b5 (Add merge-subtree back,
2008-02-23) while we're at it, since 74ca5e6d7ca (Makefile: flatten
enumeration of headers, objects and programs, 2008-03-12) it's been in
the wrong place, i.e. it applies to the manually listed ones, but was
above the glob.

We need to update the CMake rule added in 7fe07275be5 (cmake:
determine list of extra built-ins dynamically, 2020-12-04) for this,
since I'm changing the name of the variable it cares about.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>