Makefile: explicitly define the targets for "make TAGS"
Rather than maintain a "git ls-files" and "find" command to find files
to generate the tags from, as we've done since [1] and [2], let's
instead use the file definitions the Makefile was already
carrying. Due to needing to find translations in source files[3] we
had most of the data we needed already.

This makes the behavior more obvious, and makes it easy to extend this
facility to e.g. do perl-filetype regex matching with etags(1) without
guessing what extensions map to which filetypes, when that's
information we've already got in the Makefile.

The resulting file list in $(FIND_SOURCE_FILES) is almost entirely the
same. There were notable omissions from the old list. We were
excluding these before, but now include them:

    command-list.h
    config-list.h
    t/lib-t6000.sh
    <everything in sha1collisiondetection/* as noted above>

When the sha1collisiondetection/ submodule was checked out[4] we
wouldn't find any files in it, because "ls-files" won't cross the
repository boundary. We could use the --recurse-submodules files to
find those, I'm just noting the change in behavior in moving over to
using the globs. We also excluded t/lib-t6000.sh with the previous
method due to an overly eager pattern match in [5], as well as the
generated headers in $(GENERATED_H).

We included these before, but now exclude them:

    t/interop/i0000-basic.sh
    t/interop/i5500-git-daemon.sh
    t/interop/i5700-protocol-transition.sh
    t/lib-httpd/apply-one-time-perl.sh
    t/lib-httpd/broken-smart-http.sh
    t/lib-httpd/error.sh
    t/lib-httpd/error-smart-http.sh
    t/lib-httpd/incomplete-body-upload-pack-v2-http.sh
    t/lib-httpd/incomplete-length-upload-pack-v2-http.sh
    t/perf/repos/inflate-repo.sh
    t/perf/repos/many-files.sh

Similar to t/lib-t6000.sh these were overly eager pattern matches,
these are really test files, but weren't matched by the exclusion glob
added in [2].

1. f81e7c626f3 (Makefile: Add TAGS and tags targets, 2006-03-18)

2. 335339758c4 (Makefile: ask "ls-files" to list source files if
   available, 2011-10-18)

3. cd5513a7168 (i18n: Makefile: "pot" target to extract messages
   marked for translation, 2011-02-22)

4. 86cfd61e6bc (sha1dc: optionally use sha1collisiondetection as a
   submodule, 2017-07-01)

5. e6fc85b11f4 (Makefile: exclude test cruft from FIND_SOURCE_FILES,
   2016-12-14)

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>