Makefile: add Perl files to make {tags,TAGS,cscope}
This seems to work well enough for all three of them to start with as
a minimal change. For ctags it'll auto-detect Perl as *.pm and *.pl,
but not the *.perl used in this repository. It still processes them
and finds symbols, but might not be doing so in the most optimal way.

Changing that requires injecting --language options in-between the
filenames, doing so would be a good follow-up change, but let's start
with relying on the auto-detection in these tools.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>