t/lib-bash.sh: stop insisting on !POSIXLY_CORRECT
If the assertion that POSIXLY_CORRECT can't be set to run the
completion (and now, prompt) tests was ever true, it doesn't seem to
be true anymore. Let's drop this condition added in
5c293a6be48 (tests: add initial bash completion tests, 2012-04-12).

Let's add a message about what mode we're running under in case this
ever cause problems. Now we'll emit a startup message similar to
lib-gettext.sh (see 5e9637c6297 (i18n: add infrastructure for
translating Git with gettext, 2011-11-18)):

    $ bash --posix t9902-completion.sh
    # lib-bash.sh: running under 5.1.8(1)-release in POSIXLY_CORRECT mode
    ok 1 - setup for __git_find_repo_path/__gitdir tests
    [...]

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>