packfile.h: rename FOR_EACH_OBJECT_{LOCAL,PROMISOR}_ONLY flags
In the preceding commit we changed the "FOR_EACH_OBJECT_LOCAL_ONLY"
flag to only be accepted by functions in packfile.h, let's rename it
and "FOR_EACH_OBJECT_PROMISOR_ONLY" to for consistency with the rest
of the "FOR_EACH_OBJECT_SKIP_*_PACKS" flags. It's now clear from the
name that these operate on whole packfiles, not individual objects.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>