revisions API: provide and use a release_revisions()
Add a release_revisions() function to go with setup_revisions() and
repo_init_revisions(). By fixing some really common memory leaks in
the "log" family of functions we can now run a couple of more tests in
t0*.sh.

There are still a lot of things in the "struct rev_info" that are not
being freed here, and many of those are hard to fix, e.g. there's
pointers there that revisions.c sometimes owns itself, or the caller
might own if they've set those members to memory that they've
allocated.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>