{branch,tag,for-each-ref}: provide and use a ref_filter_release()
Provide a ref_filter_release() function to plug memory leaks in
"branch" and "tag", and to replace an ad-hoc release in "for-each-ref"
added in a317a553b89 (builtin/for-each-ref: free filter and UNLEAK
sorting., 2021-04-25), that can now use our new function.

Since the t0041-usage.sh test made some use of "git for-each-ref" it
can be marked as passing now, but unfortunately the main "tag",
"branch" and "for-each-ref" still have various other memory leaks.

Signed-off-by: Ævar Arnfjörð Bjarmason <avarab@gmail.com>