make-install-avargit: no RUNTIME_PREFIX, Heisenbug segfaults
I had a long debugging session for what seems to be related to
RUNTIME_PREFIX. Perhaps the issue in 1f579f62a57 (make-install-avargit
+ config.mak: debug & learn to do full incremental compile,
2021-09-16) was that too.

I'm not certain, but disabling it seems to make things reliably
compile & test. I'd end up with mystery bisects that yielded e.g. the
merge of avar/fsck-error-on-completely-invalid-7 (no amount of
bisecting would get anything but a merge), and completely removing it
from series.conf would yield it in some other place. So something
something code size?

The only tracee of any kind I got (from valgrind) that seemed to be
vaguely useful was:

==3524715==
==3524715== Invalid read of size 8
==3524715==    at 0x26F76A: run_diff_index (diff-lib.c:604)
==3524715==    by 0x26FAD7: index_differs_from (diff-lib.c:664)
==3524715==    by 0x15371D: prepare_to_commit (commit.c:1012)
==3524715==    by 0x155545: cmd_commit (commit.c:1725)
==3524715==    by 0x127A01: run_builtin (git.c:461)
==3524715==    by 0x127E06: handle_builtin (git.c:715)
==3524715==    by 0x128070: run_argv (git.c:782)
==3524715==    by 0x1284F1: cmd_main (git.c:913)
==3524715==    by 0x20BA9B: main (common-main.c:52)
==3524715==  Address 0xd0 is not stack'd, malloc'd or (recently) free'd